java-8-exercises/src/main/java/com/technologyconversations/java8exercises/streams/ToUpperCase.java
package com.technologyconversations.java8exercises.streams;

import lombok.AccessLevel;
import lombok.NoArgsConstructor;

import java.util.ArrayList;
import java.util.List;
import java.util.stream.Collectors;

@NoArgsConstructor(access = AccessLevel.PRIVATE)
public class ToUpperCase {

    public static List<String> transform7(List<String> collection) {
        List<String> newCollection = new ArrayList<>();
        for (String element : collection) {
            newCollection.add(element.toUpperCase());
        }
        return newCollection;
    }

    public static List<String> transform(List<String> collection) {

        return collection.stream().map(elem -> elem.toUpperCase()).collect(Collectors.toList());
    }

}
